+++ src/app/core/config.py
@@ -15,5 +15,5 @@ MAX_CONNECTIONS_COUNT = int(os.getenv("MAX_CONNECTIONS_COUNT", 10))
 MIN_CONNECTIONS_COUNT = int(os.getenv("MIN_CONNECTIONS_COUNT", 10))
 
 ELASTIC_INDEX = "workspace"
-ELASTIC_TOTAL_SIZE = 1_000_000  # Maximum number of entries elasticsearch queries, very large to query all entries
+ELASTIC_TOTAL_SIZE = 500_000  # Maximum number of entries elasticsearch queries, very large to query all entries
 ELASTICSEARCH_TIMEOUT = int(os.getenv("ELASTICSEARCH_TIMEOUT", 20))
